Adds private key to beacon change relayer for parachain
